requirements.txt
tifffile == 2020.7.4
torch == 1.5.1
torchvision == 0.6.1
opencv-python == 4.1.2.30
numpy == 1.17.4
matplotlib == 3.1.1
openslide-python == 1.1.2
scikit-image == 0.15.0
pandas == 1.1.3
scikit-learn == 0.23.2
